Move user serching logic to separate file
